+++ src/CodeWF.Markdown.Sample/App.axaml.cs
@@ -5,6 +5,10 @@ using Avalonia.Markup.Xaml;
 using CodeWF.Markdown.Sample.ViewModels;
 using CodeWF.Markdown.Sample.Views;
 
+using Lang.Avalonia;
+using Lang.Avalonia.Resx;
+using System.Globalization;
+
 namespace CodeWF.Markdown.Sample;
 
 public partial class App : Application
@@ -12,6 +16,7 @@ public partial class App : Application
     public override void Initialize()
     {
         AvaloniaXamlLoader.Load(this);
+        I18nManager.Instance.Register(new ResxLangPlugin(), new CultureInfo("zh-CN"), out _);
     }
 
     public override void OnFrameworkInitializationCompleted()